requirements.txt
absl-py==1.0.0
atari-py==0.2.9
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==1.6.0
configparser==5.2.0
cycler==0.11.0
docker-pycreds==0.4.0
flake8==4.0.1
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
gym==0.19.0
idna==3.3
importlib-metadata==4.10.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==9.0.0
promise==2.3
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pygame==2.1.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sentry-sdk==1.5.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stable-baselines3==1.3.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
torch==1.10.1
typing-extensions==4.0.1
urllib3==1.26.8
wandb==0.12.9
Werkzeug==2.0.2
yaspin==2.1.0
zipp==3.7.0
